[vtbackend] Refactor Sequencer into SequenceBuilder to increase code share.
Also, SequenceBuilder is the way better and more intuitive name than Sequencer :)
