+++ src/vtbackend/Metrics.h
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 #pragma once
 
-#include <vtbackend/Sequencer.h> // Sequence
-
 #include <algorithm>
 #include <map>
 #include <string>
