+++ src/vtbackend/Screen.h
@@ -10,7 +10,6 @@
 #include <vtbackend/Image.h>
 #include <vtbackend/ScreenBase.h>
 #include <vtbackend/ScreenEvents.h>
-#include <vtbackend/Sequencer.h>
 #include <vtbackend/VTType.h>
 #include <vtbackend/cell/CellConcept.h>
 
@@ -39,8 +38,89 @@
 namespace vtbackend
 {
 
+class SixelImageBuilder;
 class Terminal;
 struct Settings;
+
+// {{{ TODO: move me somewhere more appropriate
+// XTSMGRAPHICS (xterm extension)
+// CSI ? Pi ; Pa ; Pv S
+namespace XtSmGraphics
+{
+    enum class Item
+    {
+        NumberOfColorRegisters = 1,
+        SixelGraphicsGeometry = 2,
+        ReGISGraphicsGeometry = 3,
+    };
+
+    enum class Action
+    {
+        Read = 1,
+        ResetToDefault = 2,
+        SetToValue = 3,
+        ReadLimit = 4
+    };
+
+    using Value = std::variant<std::monostate, unsigned, ImageSize>;
+} // namespace XtSmGraphics
+
+/// TBC - Tab Clear
+///
+/// This control function clears tab stops.
+enum class HorizontalTabClear
+{
+    /// Ps = 0 (default)
+    AllTabs,
+
+    /// Ps = 3
+    UnderCursor,
+};
+
+/// Input: CSI 16 t
+///
+///  Input: CSI 14 t (for text area size)
+///  Input: CSI 14; 2 t (for full window size)
+/// Output: CSI 14 ; width ; height ; t
+enum class RequestPixelSize // TODO: rename RequestPixelSize to RequestArea?
+{
+    CellArea,
+    TextArea,
+    WindowArea,
+};
+
+/// DECRQSS - Request Status String
+enum class RequestStatusString
+{
+    SGR,
+    DECSCL,
+    DECSCUSR,
+    DECSCA,
+    DECSTBM,
+    DECSLRM,
+    DECSLPP,
+    DECSCPP,
+    DECSNLS,
+    DECSASD,
+    DECSSDT,
+};
+
+inline std::string setDynamicColorValue(
+    RGBColor const& color) // TODO: yet another helper. maybe SemanticsUtils static class?
+{
+    auto const r = static_cast<unsigned>(static_cast<float>(color.red) / 255.0f * 0xFFFF);
+    auto const g = static_cast<unsigned>(static_cast<float>(color.green) / 255.0f * 0xFFFF);
+    auto const b = static_cast<unsigned>(static_cast<float>(color.blue) / 255.0f * 0xFFFF);
+    return fmt::format("rgb:{:04X}/{:04X}/{:04X}", r, g, b);
+}
+
+enum class ApplyResult
+{
+    Ok,
+    Invalid,
+    Unsupported,
+};
+// }}}
 
 /**
  * Terminal Screen.
@@ -594,3 +674,45 @@ inline bool Screen<Cell>::isContiguousToCurrentLine(std::string_view continuatio
 }
 
 } // namespace vtbackend
+
+// {{{ fmt formatter
+template <>
+struct fmt::formatter<vtbackend::RequestStatusString>: formatter<std::string_view>
+{
+    auto format(vtbackend::RequestStatusString value,
+                format_context& ctx) noexcept -> format_context::iterator
+    {
+        string_view name;
+        switch (value)
+        {
+            case vtbackend::RequestStatusString::SGR: name = "SGR"; break;
+            case vtbackend::RequestStatusString::DECSCL: name = "DECSCL"; break;
+            case vtbackend::RequestStatusString::DECSCUSR: name = "DECSCUSR"; break;
+            case vtbackend::RequestStatusString::DECSCA: name = "DECSCA"; break;
+            case vtbackend::RequestStatusString::DECSTBM: name = "DECSTBM"; break;
+            case vtbackend::RequestStatusString::DECSLRM: name = "DECSLRM"; break;
+            case vtbackend::RequestStatusString::DECSLPP: name = "DECSLPP"; break;
+            case vtbackend::RequestStatusString::DECSCPP: name = "DECSCPP"; break;
+            case vtbackend::RequestStatusString::DECSNLS: name = "DECSNLS"; break;
+            case vtbackend::RequestStatusString::DECSASD: name = "DECSASD"; break;
+            case vtbackend::RequestStatusString::DECSSDT: name = "DECSSDT"; break;
+        }
+        return formatter<string_view>::format(name, ctx);
+    }
+};
+
+template <>
+struct fmt::formatter<vtbackend::Sequence>
+{
+    template <typename ParseContext>
+    constexpr auto parse(ParseContext& ctx)
+    {
+        return ctx.begin();
+    }
+    template <typename FormatContext>
+    auto format(vtbackend::Sequence const& seq, FormatContext& ctx)
+    {
+        return fmt::format_to(ctx.out(), "{}", seq.text());
+    }
+};
+// }}}
